feat: move external url function to providers
