+++ packages/bridge/src/ibc/index.ts
@@ -6,10 +6,12 @@ import { IbcTransferMethod } from "@osmosis-labs/types";
 
 import { BridgeError, BridgeQuoteError } from "../errors";
 import {
+  BridgeExternalUrl,
   BridgeProvider,
   BridgeProviderContext,
   BridgeQuote,
   CosmosBridgeTransactionRequest,
+  GetBridgeExternalUrlParams,
   GetBridgeQuoteParams,
 } from "../interface";
 import { cosmosMsgOpts } from "../msg";
@@ -226,5 +228,24 @@ export class IbcBridgeProvider implements BridgeProvider {
         },
       ]);
     }
+  }
+
+  async getExternalUrl({
+    fromChain,
+    toChain,
+    fromAsset,
+    toAsset,
+  }: GetBridgeExternalUrlParams): Promise<BridgeExternalUrl | undefined> {
+    if (fromChain.chainType === "evm" || toChain.chainType === "evm") {
+      return undefined;
+    }
+
+    const url = new URL("https://geo.tfm.com/");
+    url.searchParams.set("chainFrom", fromChain.chainId);
+    url.searchParams.set("token0", fromAsset.sourceDenom);
+    url.searchParams.set("chainTo", toChain.chainId);
+    url.searchParams.set("token1", toAsset.sourceDenom);
+
+    return { urlProviderName: "TFM", url };
   }
 }
