+++ packages/bridge/src/index.ts
@@ -1,11 +1,7 @@
 export * from "./axelar";
-export * from "./axelar/external-url";
 export * from "./bridge-providers";
 export * from "./chain";
 export * from "./errors";
-export * from "./ibc/external-url";
 export * from "./interface";
 export * from "./skip";
-export * from "./skip/external-url";
 export * from "./squid";
-export * from "./squid/external-url";
